services/users/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
